Rename models package to lib package
