+++ commands/docs/gitbook/gitbookComponents_test.go
@@ -3,8 +3,8 @@ package gitbook
 import (
 	"testing"
 
-	v2 "github.com/opencontrol/compliance-masonry/models/components/versions/2_0_0"
-	"github.com/opencontrol/compliance-masonry/models/common"
+	v2 "github.com/opencontrol/compliance-masonry/lib/components/versions/2_0_0"
+	"github.com/opencontrol/compliance-masonry/lib/common"
 	"github.com/blang/semver"
 )
