+++ lib/certifications_test.go
@@ -1,4 +1,4 @@
-package models
+package lib
 
 import (
 	"path/filepath"
