+++ lib/common/references_test.go
