+++ lib/components/versions/3_0_0/component.go
@@ -2,8 +2,8 @@ package component
 
 import (
 	"github.com/blang/semver"
-	"github.com/opencontrol/compliance-masonry/models/common"
-	"github.com/opencontrol/compliance-masonry/models/components/versions/base"
+	"github.com/opencontrol/compliance-masonry/lib/common"
+	"github.com/opencontrol/compliance-masonry/lib/components/versions/base"
 )
 
 // Component struct is an individual component requiring documentation
