+++ lib/components/versions/parse.go
@@ -5,10 +5,10 @@ import (
 
 	"github.com/blang/semver"
 	"github.com/opencontrol/compliance-masonry/config"
-	v2 "github.com/opencontrol/compliance-masonry/models/components/versions/2_0_0"
-	v3 "github.com/opencontrol/compliance-masonry/models/components/versions/3_0_0"
-	v31 "github.com/opencontrol/compliance-masonry/models/components/versions/3_1_0"
-	"github.com/opencontrol/compliance-masonry/models/components/versions/base"
+	v2 "github.com/opencontrol/compliance-masonry/lib/components/versions/2_0_0"
+	v3 "github.com/opencontrol/compliance-masonry/lib/components/versions/3_0_0"
+	v31 "github.com/opencontrol/compliance-masonry/lib/components/versions/3_1_0"
+	"github.com/opencontrol/compliance-masonry/lib/components/versions/base"
 	"gopkg.in/yaml.v2"
 )
